Change the way parse object is generated
We now create an object, with the boss names as keys, and all classes held within.

Previously we just assumed that, as we went through the list of parses, if we hit a DIFFERENT encounterID, that it was the NEXT one. But found a case where it went 1->2->1->3 which completely destroyed things.

This new way is also far simpler logic anyway.